dynamic routes + port communication hooks based on window id
