+++ src/renderer/screens/Import/index.tsx
@@ -7,10 +7,14 @@ import { ImportLedger } from './Ledger';
 import { ImportVault } from './Vault';
 import { ModalSection } from '@/renderer/kits/Overlay/structure/ModalSection';
 import { ModalMotionTwoSection } from '@/renderer/kits/Overlay/structure/ModalMotionTwoSection';
-import type { AccountSource } from '@/types/accounts';
 import { ImportReadOnly } from './ReadOnly';
+import { useImportMessagePorts } from '@/renderer/hooks/useImportMessagePorts';
+import type { AccountSource } from '@/types/accounts';
 
 export const Import: React.FC = () => {
+  // Set up port communication for `import` window.
+  useImportMessagePorts();
+
   const [source, setSource] = useState<AccountSource | undefined>('ledger');
 
   // active section
